tests: add unit tests to ShellAliasSource
